Refactor config classes
