+++ diag/src/main/scala/diagnostics/metrics/Config.scala
@@ -1,7 +1,6 @@
 package diagnostics.metrics
 
-import sdk.config.Description
-
+import sdk.reflect.Description
 import scala.concurrent.duration.{Duration, DurationInt}
 
 @Description("influxDb")
