+++ sim/src/test/scala/ReferenceConf.scala
@@ -1,10 +1,14 @@
 import org.scalatest.flatspec.AnyFlatSpec
-import sdk.config.ConfigRender
+import sdk.reflect.ClassesAsConfig
 
 class ReferenceConf extends AnyFlatSpec {
   "this" should "output reference configuration for simulation" in {
-    val s =
-      ConfigRender.referenceConf("gorki", sim.Config.Default, node.Config.Default, diagnostics.metrics.Config.Default)
+    val s = ClassesAsConfig(
+      "gorki",
+      sim.Config.Default,
+      node.Config.Default,
+      diagnostics.metrics.Config.Default,
+    )
     println(s)
   }
 }
